BasicNBodySim.jl Simulation Functions

Simulate

BasicNBodySim.simulate — Method

simulate(all_bodies::Vector{Body}, time::Float64, steps::Int64)

Simulate the gravitational force between bodies in all_bodies for time seconds using Verlet Velocity symplectic integration.

The delta time is determined by time / steps.

Update

BasicNBodySim.update! — Method

update!(current_body::Body, all_bodies::Vector{Body}, dtime::Float64)

Update the forces, positions, and velocities of the given current_body.

Get axis length

BasicNBodySim.get_axlen — Method

get_axlen(all_bodies::Vector{Body})

Get the proper axis length given the positions of each simulated body.

This is only useful when making animations with CairoMakie.jl.